Separate multiscale
